Massive code and organization update to make the program run faster

+++ CrossBoa/CrossBoa/Managers/Camera.cs
@@ -120,7 +120,7 @@ namespace CrossBoa
             {
                 cameraCrew.ApplyForce(Helper.DirectionBetween(new Point(cameraCrew.Rectangle.Center.X, 0), new Point((int)player.Position.X, 0)), Helper.DistanceSquared(new Point(cameraCrew.Rectangle.Center.X, 0), new Point((int)player.Position.X, 0)) / 1000);
                 cameraY = 0;
-                cameraX = -(int)cameraCrew.Rectangle.X;
+                cameraX = -cameraCrew.Rectangle.X;
             }
             else
             {
